+++ sentry-kotlin-compiler-plugin/src/main/kotlin/io/sentry/SentryKotlinCompilerPlugin.kt
@@ -26,15 +26,15 @@ class SentryKotlinCompilerPlugin : CompilerPluginRegistrar() {
     val versionString = KotlinCompilerVersion.getVersion()
     val version =
       if (versionString != null) {
-        SimpleSemanticVersion.from(versionString)
+        SimpleSemVer.parse(versionString)
       } else {
-        SimpleSemanticVersion(2, 1, 20)
+        SimpleSemVer(2, 1, 20)
       }
 
     val extension: IrGenerationExtension =
-      if (version >= SimpleSemanticVersion(2, 2, 0)) {
+      if (version >= SimpleSemVer(2, 2, 0)) {
         JetpackComposeTracingIrExtension22(messageCollector)
-      } else if (version >= SimpleSemanticVersion(2, 1, 20)) {
+      } else if (version >= SimpleSemVer(2, 1, 20)) {
         // 2.1.20 removed some optional parameters, causing API incompatibility
         // e.g. java.lang.NoSuchMethodError
         // see https://github.com/JetBrains/kotlin/commit/dd508452c414a0ee8082aa6f76d664271cb38f2f
@@ -44,24 +44,5 @@ class SentryKotlinCompilerPlugin : CompilerPluginRegistrar() {
       }
 
     IrGenerationExtension.registerExtension(extension)
-  }
-
-  data class SimpleSemanticVersion(val major: Int, val minor: Int, val patch: Int) :
-    Comparable<SimpleSemanticVersion> {
-
-    companion object {
-      fun from(version: String): SimpleSemanticVersion {
-        val parts = version.trim().split(".")
-        require(parts.size == 3) { "Invalid semantic version: $version" }
-
-        val (major, minor, patch) =
-          parts.map { it.toIntOrNull() ?: error("Invalid number in version: $version") }
-        return SimpleSemanticVersion(major, minor, patch)
-      }
-    }
-
-    override fun compareTo(other: SimpleSemanticVersion): Int {
-      return compareValuesBy(this, other, { it.major }, { it.minor }, { it.patch })
-    }
   }
 }